requirements.txt
altgraph==0.17.3
pyinstaller==5.6.2
pyinstaller-hooks-contrib==2022.11
python-magic==0.4.27
tabulate==0.9.0